+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # @gilamran/tsc-watch CHANGELOG
+## v7.2.0 - 29/09/2025
+
+- feature: Fixed support for tsgo (native compilation) using `--compileCommand` argument.
+
+# @gilamran/tsc-watch CHANGELOG
 ## v7.1.1 - 28/05/2025
 
 - feature: Added support for tsgo (native compilation))